Updated event binding for each content piece.
Now each content piece triggers an event with its id.  This allows
other frameworks, e.g. Backbone, to bind the element to the View when
the element is added to the DOM, without tying TimelineJS to a specific
framework, e.g. Backbone.  Keeps the code modular, only relying on
jQuery to trigger the events.